4th Revision
Started Roster Remove Function.
Added a Loop to Remove function that will iterate forward all the student objects in the array, closing the gap and moving the newly deleted student object / nullptr location to the end of the array.

Started the Student Class Student Constructor that has all the parameters.