Revision 4, v3
Invalid Email Check Started.
Currently have checking for Spaces and if the first or last character of the email address contains a period.

Still need to check that only one at (@) symbol is in the email address. Also need to verify that there are not two or more periods (.. or ...) next to each other.
I would also like to update the error message print out to be more of a full list print out instead of a loop iteration print out.

might have to look into try/catch and some type of id/error message assignment to array and then print the array upon completion.