Revision 6, v1.0
Roster Class initialization now calculates the total number of elements in the Student Data Input table and sets the numStudents variable accordingly.

Roster Class Function for Print Average days in course now course now counts the number of courses in the loop, performs a type cast on that int number to double, and divides int day sum with the double number of courses to get a double average.
NOTE:: Still need to find a way to get the size of the array for the for-loop or change the full code to allow for dynamic max array size calculation for the days in course1, ..2, ..3, etc.