Revision: 11 Version 1.0
Updated Roster Class.

Added a Parse function that takes a single parameter string input. This will be called in a for loop from main, passing each element one by one. Parseing function will split the element string up and assign them to an array that will then be passed into the add function to be have new student objects created.

Added A Roster constructor that takes a single parameter arguement for int size of the input studentdata. This will dynamically create the pointer array based on this input.

Added to the main function the new parse function and for loop. Moved the array size calculation function above the roster class/object call to pass the input to the the constructor.

Updated Roster header file accordingly.