Revision: 12 Version: 1.0
Made two new function in Roster.
Get Input Size and a new Parse Input function.

GetInputSize function calls the studentData table array inside the roster class and calculates its size and returns an Int. This int is then used to create the size of the array and set the loop variables.

The new Parse Input function does the same as the last one but this time, instead of receiving the studentData table's elements one at a time from a loop in main. This one takes the studentData table from the roster class, uses the inputsize int returned above, and performs the original loop I had wanted to use before but now inside the class constructor.