Version: 3.5
Possible Bug was found in Remove Function. For loop responsible for moving the deleted nullptr to the end of the array ran every time. This has been corrected by adding an If statement to validate true with the isFound flag from the previous ID match/delete student loop in the same function.

ToDo:
- Code Refactor
- Submission.